Coordinator dashboard validating new users

+++ internships/src/main/java/br/com/uff/internships/repository/AbstractDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.NoRepositoryBean;
 public interface AbstractDAO<T> {
 	
 	public void create(T entity);
-	public T find(Long id);
+	public T find(Integer id);
 	public void update(T entity);
 	public void remove (T entity);
 }
